@@ -2,10 +2,10 @@
 	"name": "pax.libx.js",
 	"version": "0.2.5",
 	"description": "Flexible web assets packing and bundling library",
-	"main": "src/index.js",
 	"repository": "https://github.com/Livshitz/pax.libx.js",
 	"author": "Eli Livshitz",
 	"license": "MIT",
+	"main": "build/index.js",
 	"types": "types/pax.libx.d.ts",
 	"scripts": {
 		"build": "tsc",
@@ -20,61 +20,65 @@
 		"./build"
 	],
 	"dependencies": {
-		"@babel/cli": "^7.4.4",
-		"@babel/core": "^7.4.4",
-		"@babel/plugin-transform-runtime": "^7.4.4",
-		"@babel/preset-env": "^7.4.4",
-		"@babel/preset-react": "^7.0.0",
-		"@babel/preset-typescript": "^7.3.3",
-		"@types/jest": "^24.0.12",
-		"babel-jest": "^24.8.0",
+		"@babel/cli": "^7.11.6",
+		"@babel/core": "^7.11.6",
+		"@babel/plugin-proposal-decorators": "^7.10.5",
+		"@babel/plugin-transform-runtime": "^7.11.5",
+		"@babel/preset-env": "^7.11.5",
+		"@babel/preset-react": "^7.10.4",
+		"@babel/preset-typescript": "^7.10.4",
+		"@types/jest": "^26.0.14",
+		"babel-jest": "^26.3.0",
 		"babel-plugin-transform-es2015-arrow-functions": "^6.22.0",
 		"babel-preset-es3": "^1.0.1",
 		"babelify": "^10.0.0",
-		"browserify": "^16.2.3",
-		"cheerio": "^1.0.0-rc.2",
-		"chokidar": "^2.0.4",
-		"compression": "^1.7.3",
+		"browserify": "^16.5.2",
+		"cheerio": "^1.0.0-rc.3",
+		"chokidar": "^3.4.2",
+		"compression": "^1.7.4",
 		"cors": "^2.8.5",
-		"del": "^3.0.0",
-		"gulp": "^3.9.1",
-		"gulp-angular-templatecache": "^2.2.6",
+		"del": "^5.1.0",
+		"esmify": "^2.1.1",
+		"gulp": "^4.0.2",
+		"gulp-angular-templatecache": "^3.0.0",
 		"gulp-babel": "^8.0.0",
-		"gulp-babel-minify": "^0.5.0",
-		"gulp-clean-css": "^4.0.0",
+		"gulp-babel-minify": "^0.5.1",
+		"gulp-clean-css": "^4.3.0",
 		"gulp-concat": "^2.6.1",
 		"gulp-connect": "^5.7.0",
 		"gulp-debug": "^4.0.0",
 		"gulp-htmlmin": "^5.0.1",
-		"gulp-if": "^2.0.2",
+		"gulp-if": "^3.0.0",
 		"gulp-less": "^4.0.1",
 		"gulp-less2sass": "^0.0.2",
 		"gulp-pug": "^4.0.1",
-		"gulp-rename": "1.2.2",
-		"gulp-sass": "^4.0.2",
-		"gulp-shell": "^0.6.5",
-		"gulp-sourcemaps": "^2.6.4",
+		"gulp-rename": "2.0.0",
+		"gulp-sass": "^4.1.0",
+		"gulp-shell": "^0.8.0",
+		"gulp-sourcemaps": "^2.6.5",
 		"gulp-streamify": "^1.0.2",
-		"gulp-typescript": "^5.0.1",
-		"gulp-uglify": "^3.0.1",
-		"gulp-usemin": "^0.3.29",
+		"gulp-typescript": "^6.0.0-alpha.1",
+		"gulp-uglify": "^3.0.2",
+		"gulp-usemin": "^0.3.30",
 		"gulp-watch": "^5.0.1",
-		"into-stream": "^4.0.0",
-		"jest": "^24.8.0",
+		"into-stream": "^6.0.0",
+		"jest": "^26.4.2",
+		"less": "^3.12.2",
 		"less-plugin-sass2less": "^1.2.0",
 		"libx.js": "*",
 		"node-sass": "^4.14.1",
-		"pug": "^2.0.4",
-		"readable-stream": "^3.3.0",
+		"pug": "^3.0.0",
+		"readable-stream": "^3.6.0",
 		"through": "^2.3.8",
-		"through2": "^3.0.0",
-		"ts-jest": "^24.0.2",
-		"tsify": "^4.0.1",
-		"typescript": "^3.4.5",
+		"through2": "^4.0.2",
+		"ts-jest": "^26.3.0",
+		"tsify": "^5.0.2",
+		"typescript": "^4.0.2",
+		"vinyl": "^2.2.1",
 		"vinyl-buffer": "^1.0.1",
 		"vinyl-source-stream": "^2.0.0",
 		"vinyl-transform": "^1.0.0",
-		"vue-template-compiler": "^2.6.11"
+		"vue-template-compiler": "^2.6.12"
 	},
 	"not-in-use-dependencies": {
 		"connect-gzip": "^0.1.6",
@@ -89,7 +93,7 @@
 		]
 	},
 	"devDependencies": {
-		"jest-junit": "^10.0.0",
+		"jest-junit": "^11.1.0",
 		"jest-junit-reporter": "^1.1.0"
 	},
 	"jest": {
