@@ -3,6 +3,6 @@ let bundler = require('./bundler');
 
 export default module.exports = bundler;
 
-(()=>{ // Dependency Injector auto module registration 
+(()=>{ // Dependency Injector auto module registration   
 	libx.di.register('pax', module.exports);
 })();
